Retour Coupe 2011

+++ Dev_uCOS-II/Oufff2011/src/Main/LibMoving.h
@@ -21,14 +21,14 @@ void LibMoving_MoveInMM(int Dist, INT8U Speed, StructCmd *NextSetpoint);
 	// MvtMode is set to 'MvtMode_Simple'
 	// CmdType is set to 'CmdType_Blocking'
 
-void LibMoving_RotateInDeg(float AngleInDeg, INT8U Speed, StructCmd *NextSetpoint);
+void LibMoving_RotateInDeg(int AngleInDeg, INT8U Speed, StructCmd *NextSetpoint);
 	// AngleInDeg   : in mm for the movement (>0 go forward, <0 go back)
     // Speed        : Speed to use 
 	// NextSetpoint	: pointer to the struct for containing new position
 	// MvtMode is set to 'MvtMode_Simple'
 	// CmdType is set to 'CmdType_Blocking'
 
-void LibMoving_MoveToAngleInDeg(float AngleToGoInDeg, INT8U Speed, StructCmd *NextSetpoint);
+void LibMoving_MoveToAngleInDeg(int AngleToGoInDeg, INT8U Speed, StructCmd *NextSetpoint);
 	// AngleToGoInDeg	: angle in degree we want to go
     // Speed            : Speed to use 
 	// NextSetpoint	    : pointer to the struct for containing new position
